startpoint/requirements.txt
pytest
pytest-docker-pexpect
pytest-ordering
pytest-logger
defusedxml
PySide2